#38: Starting again activity does does not use category of past activity
This was possibly a problem with hamster legacy. In the gk+2 world,
starting a fact with an empty category via dbus would autmatically
append the most recent cateogory of the fact. Now with gtk+3,
an empty category is passed through verbatim.

Thankfully, filtering the GetActivities method from the entry text
returns the categories in recent order and we can pick the last used.